+++ snapshot-tests-core/src/test/java/de/skuzzle/test/snapshots/SnapshotsTest.java
@@ -25,6 +25,8 @@ public class SnapshotsTest {
 
         assertThat(testResult.rawActualResultFile()).exists();
         assertThat(testResult.actualResultFile()).exists();
+        assertThat(testResult.contextFiles().rawActualResultFile()).exists();
+        assertThat(testResult.contextFiles().actualResultFile()).exists();
     }
 
     @Test
@@ -43,6 +45,8 @@ public class SnapshotsTest {
 
         assertThat(testResult.rawActualResultFile()).exists();
         assertThat(testResult.actualResultFile()).exists();
+        assertThat(testResult.contextFiles().rawActualResultFile()).exists();
+        assertThat(testResult.contextFiles().actualResultFile()).exists();
     }
 
     @Test
@@ -55,6 +59,8 @@ public class SnapshotsTest {
 
         assertThat(testResult.rawActualResultFile()).exists();
         assertThat(testResult.actualResultFile()).exists();
+        assertThat(testResult.contextFiles().rawActualResultFile()).exists();
+        assertThat(testResult.contextFiles().actualResultFile()).exists();
     }
 
     @Test
@@ -64,6 +70,7 @@ public class SnapshotsTest {
 
         assertThat(testResultDisabled.status()).isEqualTo(SnapshotStatus.DISABLED);
         assertThat(testResultDisabled.targetFile()).doesNotExist();
+        assertThat(testResultDisabled.contextFiles().snapshotFile()).doesNotExist();
 
         final Person phil = determinePerson().setName("Phil");
         final SnapshotTestResult testResultActive = snapshot.assertThat(phil).asText().matchesSnapshotText();
@@ -78,6 +85,7 @@ public class SnapshotsTest {
 
         assertThat(testResultDisabled.status()).isEqualTo(SnapshotStatus.DISABLED);
         assertThat(testResultDisabled.targetFile()).exists();
+        assertThat(testResultDisabled.contextFiles().snapshotFile()).exists();
 
         final Person phil = determinePerson().setName("Phil");
         snapshot.assertThat(phil).asText().matchesSnapshotText();