+++ snapshot-tests-core/src/test/java/de/skuzzle/test/snapshots/impl/FailingSnapshotTests.java
@@ -33,11 +33,27 @@ public class FailingSnapshotTests {
     static class DetectIncompleteDslReuse {
 
         @Test
-        void testIllegalReuse(Snapshot snapshot) throws Exception {
+        void testIllegalReuseOfAssertThat(Snapshot snapshot) throws Exception {
             MetaTest.assumeMetaTest();
 
             snapshot.assertThat("");
             snapshot.assertThat("");
+        }
+
+        @Test
+        void testIllegalReuseOfNamed(Snapshot snapshot) throws Exception {
+            MetaTest.assumeMetaTest();
+
+            snapshot.named("");
+            snapshot.named("");
+        }
+
+        @Test
+        void testIllegalReuseOfIn(Snapshot snapshot) throws Exception {
+            MetaTest.assumeMetaTest();
+
+            snapshot.in(Paths.get("egal"));
+            snapshot.in(Paths.get("egal"));
         }
     }
 
@@ -89,7 +105,7 @@ public class FailingSnapshotTests {
         frameworkTest.expectTestcase(FailBecauseOfNullInputSnapshotAlreadyExists.class)
                 .toFailWithExceptionWhich()
                 .isInstanceOf(AssertionError.class)
-                .hasMessage("Expected actual not to be null but to match stored snapshot:\n\nsnapshot text");
+                .hasMessage("Expected actual not to be null in order to take snapshot");
     }
 
     @EnableSnapshotTests
@@ -108,7 +124,7 @@ public class FailingSnapshotTests {
         frameworkTest.expectTestcase(FailBecauseOfNullInputJustUpdateSnapshot.class)
                 .toFailWithExceptionWhich()
                 .isInstanceOf(AssertionError.class)
-                .hasMessage("Expected actual not to be null in order to take initial snapshot");
+                .hasMessage("Expected actual not to be null in order to take snapshot");
     }
 
     @EnableSnapshotTests
@@ -127,7 +143,7 @@ public class FailingSnapshotTests {
         frameworkTest.expectTestcase(FailBecauseOfNullInputInitialSnapshot.class)
                 .toFailWithExceptionWhich()
                 .isInstanceOf(AssertionError.class)
-                .hasMessage("Expected actual not to be null in order to take initial snapshot");
+                .hasMessage("Expected actual not to be null in order to take snapshot");
     }
 
     @EnableSnapshotTests
@@ -148,7 +164,7 @@ public class FailingSnapshotTests {
                 .toFailWithExceptionWhich()
                 .isInstanceOf(AssertionError.class)
                 .hasMessage(String.format(
-                        "Snapshots have been updated forcefully.%nRemove 'updateSnapshots = true' attribute from your test class and calls to 'justUpdateSnapshot()' and run the tests again."));
+                        "Snapshots have been updated forcefully.%nRemove '@ForceUpdateSnapshots' annotation from your test class and calls to 'justUpdateSnapshot()' then run the tests again."));
     }
 
     @EnableSnapshotTests(forceUpdateSnapshots = true) // leave force true
@@ -170,7 +186,7 @@ public class FailingSnapshotTests {
                 .toFailWithExceptionWhich()
                 .isInstanceOf(AssertionError.class)
                 .hasMessage(String.format(
-                        "Snapshots have been updated forcefully.%nRemove 'updateSnapshots = true' attribute from your test class and calls to 'justUpdateSnapshot()' and run the tests again."));
+                        "Snapshots have been updated forcefully.%nRemove '@ForceUpdateSnapshots' annotation from your test class and calls to 'justUpdateSnapshot()' then run the tests again."));
     }
 
     @EnableSnapshotTests
@@ -194,7 +210,7 @@ public class FailingSnapshotTests {
                 .isInstanceOf(AssertionError.class)
                 .hasMessage(String.format(
                         "Snapshots have been updated forcefully.%n"
-                                + "Remove 'updateSnapshots = true' attribute from your test class and calls to 'justUpdateSnapshot()' and run the tests again."));
+                                + "Remove '@ForceUpdateSnapshots' annotation from your test class and calls to 'justUpdateSnapshot()' then run the tests again."));
     }
 
     @EnableSnapshotTests
@@ -216,7 +232,7 @@ public class FailingSnapshotTests {
                 .isInstanceOf(AssertionError.class)
                 .hasMessage(String.format(
                         "Snapshots have been updated forcefully.%n"
-                                + "Remove 'updateSnapshots = true' attribute from your test class and calls to 'justUpdateSnapshot()' and run the tests again."));
+                                + "Remove '@ForceUpdateSnapshots' annotation from your test class and calls to 'justUpdateSnapshot()' then run the tests again."));
     }
 
     @EnableSnapshotTests